+++ lib/platform/io.dart
@@ -1,20 +1,24 @@
 import 'dart:async';
-import 'package:distributed/interfaces/node.dart';
+
 import 'package:distributed/src/configuration.dart';
 import 'package:distributed/src/io/node.dart';
 import 'package:seltzer/platform/server.dart';
 
+export 'package:distributed/src/io/node_repl.dart' show NodeREPL;
+
 void configureDistributed() {
   useSeltzerInTheServer();
-  setNodeProvider(_createNode);
+  setNodeProvider(new IONodeProvider());
 }
 
-/// Creates a new [Node] identified by [name] and [hostname].
-Future<IONode> _createNode(String name, String hostname, String cookie,
-        {int port: 9095, bool hidden: false}) =>
-    IONode.create(
-        name: name,
-        hostname: hostname,
-        port: port,
-        cookie: cookie,
-        isHidden: hidden);
+class IONodeProvider implements NodeProvider {
+  @override
+  Future<IONode> create(String name, String hostname, String cookie,
+          {int port: 9095, bool hidden: false}) =>
+      IONode.create(
+          name: name,
+          hostname: hostname,
+          port: port,
+          cookie: cookie,
+          isHidden: hidden);
+}
