+++ lib/src/configuration.dart
@@ -1,8 +1,12 @@
 import 'dart:async';
+
 import 'package:distributed/interfaces/node.dart';
 
-typedef Future<Node> NodeProvider(String name, String hostname, String cookie,
-    {int port, bool hidden});
+abstract class NodeProvider {
+  /// Creates a new [Node] identified by [name] and [hostname].
+  Future<Node> create(String name, String hostname, String cookie,
+      {int port, bool hidden});
+}
 
 NodeProvider _nodeProvider;
 
@@ -15,4 +19,4 @@ void setNodeProvider(NodeProvider nodeProvider) {
 
 Future<Node> createNode(String name, String hostname, String cookie,
         {int port, bool hidden}) =>
-    _nodeProvider(name, hostname, cookie, port: port, hidden: hidden);
+    _nodeProvider.create(name, hostname, cookie, port: port, hidden: hidden);
